Analyze bug in blaze utf8 renderer
